Several changes including formatting and softwares
The formatting has been updated to comply with puppet lint checks.
The softwares is no longer a yaml file located in the module directory,
instead its a parameter called softwares that is a hash which is written
to the softwares.yaml file via templates.
The install exec command no longer runs on every puppet run but only
runs when the softwares file has changed.